Move things around a little more, keep trying to improve docs
