+++ tests/custom_db_mapping.py
@@ -1,0 +1,117 @@
+######################################################################################################################
+# Copyright (C) 2017-2022 Spine project consortium
+# This file is part of Spine Database API.
+# Spine Database API is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser
+# General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your
+# option) any later version. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+# without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General
+# Public License for more details. You should have received a copy of the GNU Lesser General Public License along with
+# this program. If not, see <http://www.gnu.org/licenses/>.
+######################################################################################################################
+
+"""
+Unit tests for DatabaseMapping class.
+
+"""
+from spinedb_api import DatabaseMapping, SpineIntegrityError
+
+
+class CustomDatabaseMapping(DatabaseMapping):
+    def add_object_classes(self, *items, **kwargs):
+        return self.add_items("object_class", *items, **kwargs)
+
+    def add_objects(self, *items, **kwargs):
+        return self.add_items("object", *items, **kwargs)
+
+    def add_entity_classes(self, *items, **kwargs):
+        return self.add_items("entity_class", *items, **kwargs)
+
+    def add_entities(self, *items, **kwargs):
+        return self.add_items("entity", *items, **kwargs)
+
+    def add_wide_relationship_classes(self, *items, **kwargs):
+        return self.add_items("relationship_class", *items, **kwargs)
+
+    def add_wide_relationships(self, *items, **kwargs):
+        return self.add_items("relationship", *items, **kwargs)
+
+    def add_parameter_definitions(self, *items, **kwargs):
+        return self.add_items("parameter_definition", *items, **kwargs)
+
+    def add_parameter_values(self, *items, **kwargs):
+        return self.add_items("parameter_value", *items, **kwargs)
+
+    def add_parameter_value_lists(self, *items, **kwargs):
+        return self.add_items("parameter_value_list", *items, **kwargs)
+
+    def add_list_values(self, *items, **kwargs):
+        return self.add_items("list_value", *items, **kwargs)
+
+    def add_alternatives(self, *items, **kwargs):
+        return self.add_items("alternative", *items, **kwargs)
+
+    def add_scenarios(self, *items, **kwargs):
+        return self.add_items("scenario", *items, **kwargs)
+
+    def add_scenario_alternatives(self, *items, **kwargs):
+        return self.add_items("scenario_alternative", *items, **kwargs)
+
+    def add_entity_groups(self, *items, **kwargs):
+        return self.add_items("entity_group", *items, **kwargs)
+
+    def add_metadata(self, *items, **kwargs):
+        return self.add_items("metadata", *items, **kwargs)
+
+    def add_entity_metadata(self, *items, **kwargs):
+        return self.add_items("entity_metadata", *items, **kwargs)
+
+    def add_parameter_value_metadata(self, *items, **kwargs):
+        return self.add_items("parameter_value_metadata", *items, **kwargs)
+
+    def update_alternatives(self, *items, **kwargs):
+        return self.update_items("alternative", *items, **kwargs)
+
+    def update_scenarios(self, *items, **kwargs):
+        return self.update_items("scenario", *items, **kwargs)
+
+    def update_scenario_alternatives(self, *items, **kwargs):
+        return self.update_items("scenario_alternative", *items, **kwargs)
+
+    def update_entity_classes(self, *items, **kwargs):
+        return self.update_items("entity_class", *items, **kwargs)
+
+    def update_entities(self, *items, **kwargs):
+        return self.update_items("entity", *items, **kwargs)
+
+    def update_object_classes(self, *items, **kwargs):
+        return self.update_items("object_class", *items, **kwargs)
+
+    def update_objects(self, *items, **kwargs):
+        return self.update_items("object", *items, **kwargs)
+
+    def update_wide_relationship_classes(self, *items, **kwargs):
+        return self.update_items("relationship_class", *items, **kwargs)
+
+    def update_wide_relationships(self, *items, **kwargs):
+        return self.update_items("relationship", *items, **kwargs)
+
+    def update_parameter_definitions(self, *items, **kwargs):
+        return self.update_items("parameter_definition", *items, **kwargs)
+
+    def update_parameter_values(self, *items, **kwargs):
+        return self.update_items("parameter_value", *items, **kwargs)
+
+    def update_parameter_value_lists(self, *items, **kwargs):
+        return self.update_items("parameter_value_list", *items, **kwargs)
+
+    def update_list_values(self, *items, **kwargs):
+        return self.update_items("list_value", *items, **kwargs)
+
+    def update_metadata(self, *items, **kwargs):
+        return self.update_items("metadata", *items, **kwargs)
+
+    def update_entity_metadata(self, *items, **kwargs):
+        return self.update_items("entity_metadata", *items, **kwargs)
+
+    def update_parameter_value_metadata(self, *items, **kwargs):
+        return self.update_items("parameter_value_metadata", *items, **kwargs)
